% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{add_pathlos_data}
\alias{add_pathlos_data}
\title{Add multi-state path and length of stay data at interim analysis time tp}
\usage{
add_pathlos_data(dat, tp = NULL, delt = NULL)
}
\arguments{
\item{dat}{data.frame of class mpti_weib_dat or impti_weib_dat}

\item{tp}{interim analysis trial time}

\item{delt}{time lag of minimal follow-up required to be "enrolled"}
}
\value{
data.frame of class impti_weib_dat
}
\description{
Add multi-state path and length of stay data at interim analysis time tp
}
