% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comp_pp_eos}
\alias{comp_pp_eos}
\title{Compute posterior predictive trial summaries}
\usage{
comp_pp_eos(x, threshold = list(p = 0.05, hr = 0.8))
}
\arguments{
\item{x}{object of class pp_eos}

\item{threshold}{list of thresholds}
}
\value{
list of summaries
}
\description{
Compute posterior predictive trial summaries
}
