% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\name{create_mstate}
\alias{create_mstate}
\title{Create a Multistate data class Object}
\usage{
create_mstate(
  data,
  resp,
  resp_levels,
  surv,
  surv_evnt = NA,
  surv_csnr = NA,
  pfs,
  pfs_evnt = NA,
  pfs_csnr = NA,
  accT,
  id = NA,
  trt = NA,
  cntrl = NA,
  exper = NA,
  resp_t = NA,
  covar = NA,
  ttp = NA,
  ttp_evnt = NA,
  ttp_csnr = NA,
  cycle_dur_weeks = 4,
  tunits = "months",
  accTime = NULL,
  accExN = NULL,
  addFUt = NULL
)
}
\arguments{
\item{data}{data.frame containing patient-level outcomes and treatment}

\item{resp}{name of the surrogate response variable (var is factor)}

\item{resp_levels}{vector of stratum of resp variable that indicate a "response" (e.g. RECIST PR or CR)}

\item{surv}{name of overall survival duration variable (var is numeric)}

\item{surv_evnt}{name of overall survival event indicator (integer 0-1)}

\item{surv_csnr}{name of overall survival right-censoring indicator (integer 0-1)}

\item{pfs}{name of progression-free survival duration variable (var is numeric)}

\item{pfs_evnt}{name of progression-free survival event indicator (integer 0-1)}

\item{pfs_csnr}{name of progression-free survival right-censoring indicator (integer 0-1)}

\item{accT}{name of the duration since study start to accrual (var must be positive numeric value)}

\item{id}{name of the unique subject identifier}

\item{trt}{name of the treatment indicator variable (var can be integer or factor)}

\item{cntrl}{name of the control arm group (in treatment variable)}

\item{exper}{name of the experimental arm group (in treatment variable)}

\item{resp_t}{name of the variable describing the time of observed response}

\item{covar}{name of prognostic covariates}

\item{ttp}{name of time-to-progression duration variable (var is numeric; needed only if missing PFS & Survival)}

\item{ttp_evnt}{name of time-to-progression event indicator (integer 0-1; needed only if missing PFS & Survival)}

\item{ttp_csnr}{name of time-to-progression right-censoring indicator (integer 0-1; needed only if missing PFS & Survival)}

\item{cycle_dur_weeks}{duration of first cycle in weeks}

\item{tunits}{label for time units}

\item{accTime}{time parition for accrual}

\item{accExN}{expected enrollment for each bin of accTime}

\item{addFUt}{added follow-up duration in months before final analysis}
}
\value{
An object of class \code{mstate_dat}.
}
\description{
Create a new instance of the \code{mstate_dat} class.
}
