% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_a}
\alias{gen_a}
\title{Randomly generate treatment assignment for n pts}
\usage{
gen_a(n, p = NULL, seed = 1)
}
\arguments{
\item{n}{total sample size}

\item{p}{allocation probability to treatment}
}
\value{
vector of assignments
}
\description{
Randomly generate treatment assignment for n pts
}
