% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_accural}
\alias{gen_accural}
\title{Generate peicewise accural}
\usage{
gen_accural(n, accTime = 40, accExN = NULL)
}
\arguments{
\item{n}{total sample size}

\item{accTime}{time parition for simulating accrual}

\item{accExN}{expected enrollment for each bin of accTime}
}
\value{
sorted enrollment times for n patients
}
\description{
Generate peicewise accural
}
