% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_dur_weibph}
\alias{gen_dur_weibph}
\title{Randomly generate survival endpoint duration from Weibull PH model}
\usage{
gen_dur_weibph(
  n,
  a = matrix(nrow = n, ncol = 0),
  x = matrix(nrow = n, ncol = 0),
  m = matrix(nrow = n, ncol = 0),
  coef_a = 0,
  coef_x = rep(0, ncol(x)),
  coef_m = rep(0, ncol(m)),
  weibpars = list(nv = 1, lamb = 0.05)
)
}
\arguments{
\item{n}{total sample size}

\item{a}{treatment arm}

\item{x}{baseline covariates}

\item{m}{mediators}

\item{coef_a}{coef of treatment a}

\item{coef_x}{coefs of x}

\item{coef_m}{coefs of m}

\item{weibpars}{weib baseline distribution}
}
\value{
vector of endpoint durations
}
\description{
Randomly generate survival endpoint duration from Weibull PH model
}
