% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_heterogeneous_data}
\alias{gen_heterogeneous_data}
\title{Randomly generate treatment assignment for n pts}
\usage{
gen_heterogeneous_data(
  parm,
  n,
  a,
  x,
  nb,
  nc,
  prtrt = 0.5,
  accTime = 40,
  accExN = NULL,
  addFUt = 0,
  CenUpLim = NULL,
  tunits = "months",
  eos_stats = TRUE,
  seed = 1
)
}
\arguments{
\item{n}{total sample size}

\item{a}{treatment indicator}

\item{x}{covariate variables}

\item{nb}{number of binary variables in weibull model}

\item{nc}{number of continuous variables in weibull model}
}
\value{
vector of assignments
}
\description{
Randomly generate treatment assignment for n pts
}
