% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_m}
\alias{gen_m}
\title{Randomly generate independent patient mediators from binomial and normal distributions}
\usage{
gen_m(
  n,
  qb = 1,
  qc = 1,
  a = NULL,
  x = NULL,
  int = NULL,
  coef_a = NULL,
  coef_x = NULL,
  err_sigm = NULL,
  seed = 1
)
}
\arguments{
\item{n}{total sample size}

\item{qb}{number of binary mediators}

\item{qc}{number of continuous normal mediator}

\item{a}{vector of treatment arm indicators}

\item{x}{matrix of baseline covariates}

\item{int}{vector of qb+qc intercepts in linear predictor for mediator models, default 1}

\item{coef_a}{vector of qb+qc coefficients for a in linear predictor of surrogate models, default 0}

\item{coef_x}{matrix of coefficients for x in linear predictor of surrogate models, one row per mediator, default 0}

\item{err_sigm}{sd of error terms in normal mediators, length of qc, default 0.5}
}
\value{
model matrix
}
\description{
Randomly generate independent patient mediators from binomial and normal distributions
}
