% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_m_obs}
\alias{gen_m_obs}
\title{Randomly generate observation durations for mediators}
\usage{
gen_m_obs(
  n,
  q = 1,
  eos_dur = 50,
  weib_shape = NULL,
  weib_scale = NULL,
  by_visit = TRUE,
  mvis = NULL,
  fix_gapt = TRUE,
  mgapt = NULL,
  seed = 1
)
}
\arguments{
\item{n}{total sample size}

\item{q}{number of mediator that will not be observed at baseline}

\item{eos_dur}{end of study duration, used if the mediation time is larger than all visit time}

\item{weib_shape}{vector of q weibull shape pars of incidence time for mediation variables}

\item{weib_scale}{vector of q weibull scale pars of incidence time for mediation variables}

\item{by_visit}{whether to use grid of study visits for the incidence time}

\item{mvis}{mean number of visits for each patient}

\item{fix_gapt}{whether fix the gap time at mgapt, if not will be generated from exp dist with mean mgapt}

\item{mgapt}{mean gap time between two visits}
}
\value{
list of generated and observed observation durations
}
\description{
Randomly generate observation durations for mediators
}
