% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_mstate}
\alias{gen_mstate}
\title{Generate dataset from Weibull multi-state mediation model with binary surrogate (mediator)}
\usage{
gen_mstate(
  parm,
  n,
  nb = 1,
  nc = 1,
  prtrt = 0.5,
  accTime = 40,
  accExN = NULL,
  addFUt = 0,
  CenUpLim = NULL,
  tunits = "months",
  eos_stats = TRUE
)
}
\arguments{
\item{parm}{parameter list of class mpti_weib_parm}

\item{n}{total sample size}

\item{nb}{number of binary prognostic covariates}

\item{nc}{number of continous (numeric) prognostic covariates}

\item{prtrt}{prob of treatment}

\item{accTime}{time parition for simulating accrual}

\item{accExN}{expected enrollment for each bin of accTime}

\item{addFUt}{added follow-up duration in months before final analysis}

\item{CenUpLim}{upper bound of right censoring for final analysis (longest duration possible to observe)}

\item{tunits}{label for time units}

\item{eos_stats}{compute end of study stats}
}
\value{
simulated dataset data frame
}
\description{
Generate dataset from Weibull multi-state mediation model with binary surrogate (mediator)
}
