% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{gen_x}
\alias{gen_x}
\title{Randomly generate independent patient baseline/prognostic covariate from binomial and normal distributions}
\usage{
gen_x(n, pb = 1, pc = 1, prob = NULL, mu = NULL, sigma = NULL, seed = 1)
}
\arguments{
\item{n}{total sample size}

\item{pb}{number of binary covariates}

\item{pc}{number of continuous covariates}

\item{prob}{vector of probabilities for binary covariates, default 0.5}

\item{mu}{vector of mean for continuous covariates, default 0}

\item{sigma}{vector of sigma for continuous covariates, default 1}
}
\value{
model matrix
}
\description{
Randomly generate independent patient baseline/prognostic covariate from binomial and normal distributions
}
