% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\docType{class}
\name{impti_weib_dat-class}
\alias{impti_weib_dat-class}
\title{General class for Weibull Generated multistate interim data}
\description{
General class for Weibull Generated multistate interim data
}
