% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{interim_mstate.mpti_weib_dat}
\alias{interim_mstate.mpti_weib_dat}
\title{Cut multi-state data at interim analysis time tp mpti_weib_dat}
\usage{
\method{interim_mstate}{mpti_weib_dat}(dat, tp, delt = 0)
}
\arguments{
\item{dat}{data.frame of class mpti_weib_dat}

\item{tp}{interim analysis trial time}

\item{delt}{time lag of minimal follow-up required to be "enrolled"}
}
\value{
data.frame of class impti_weib_dat
}
\description{
Cut multi-state data at interim analysis time tp mpti_weib_dat
}
