% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generate.R
\name{interim_mstate}
\alias{interim_mstate}
\title{Cut multi-state data at interim analysis time tp}
\usage{
interim_mstate(dat, ...)
}
\arguments{
\item{dat}{either data.frame of class mpti_weib_dat or object of class mstate_data}

\item{...}{additional arguments to class specific function calls}
}
\description{
Cut multi-state data at interim analysis time tp
}
