% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\docType{class}
\name{mstate_data-class}
\alias{mstate_data-class}
\title{General Multistate data class}
\description{
A class representing a multistate data objects not simulated from a multistate model
}
\section{Slots}{

\describe{
\item{\code{data}}{data.frame of class mpti_weib_dat containing patient-level outcomes and treatment}

\item{\code{cntrl}}{name of the unique subject identifier}

\item{\code{exper}}{name of the duration since study start to accrual (var must be positive numeric value)}

\item{\code{resp_levels}}{vector of stratum of resp variable that indicate a "response" (e.g. RECIST PR or CR)}

\item{\code{resp_t}}{logical if variable name of duration from start of study to "response" is included}

\item{\code{covar}}{name of prognostic covariates}

\item{\code{ttp}}{logical if time to progression duration variable is included}

\item{\code{dor}}{logical if duration of response variable is included}
}}

