% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{parm_mpti_weib}
\alias{parm_mpti_weib}
\title{Specify Weibull MPTI model parameters for binary surrogate}
\usage{
parm_mpti_weib(
  beta,
  gamma,
  weib_nv,
  weib_lamb,
  prob = NULL,
  mu = NULL,
  sigma = NULL,
  weib_shape = NULL,
  weib_scale = NULL,
  scenario = NULL
)
}
\arguments{
\item{beta}{coefficients for treatment, surrogate response, and prognostic factors (x) in each multistate model}

\item{gamma}{vector of intercepts and regression coefficients for treatment and baseline covariates x in surrogate response model}

\item{weib_nv}{baseline hazard Weibull shape parameters for PFS (12), OS (13), and OS (23) after progression}

\item{weib_lamb}{baseline hazard Weibull rate parameters for PFS (12), OS (13), and OS (23) after progression}

\item{prob}{binomial probability vector for binary prognostic covariates}

\item{mu}{mean vector for numeric prognostic covariates}

\item{sigma}{stand dev vector for numeric prognostic covariates}

\item{weib_shape}{Weibull shape paramter for time-to-observing surrogate response}

\item{weib_scale}{Weibull scale paramter for time-to-observing surrogate response}

\item{scenario}{text string to label the scenario}
}
\value{
list of model parameters
}
\description{
Specify Weibull MPTI model parameters for binary surrogate
}
