% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pp_eos}
\alias{plot_pp_eos}
\title{Plot Survival end of study posterior predictive trial summaries}
\usage{
plot_pp_eos(x, threshold = list(p = 0.05, hr = 0.8))
}
\arguments{
\item{x}{object of class pp_eos}

\item{threshold}{list of thresholds}
}
\value{
ggplot2 object
}
\description{
Plot Survival end of study posterior predictive trial summaries
}
