% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rr}
\alias{plot_rr}
\title{Plotting function for survival mediation modeling}
\usage{
plot_rr(x, type = "riskratio")
}
\arguments{
\item{x}{list of class riskratio}

\item{type}{text string indicating type of plot "riskratio", "survival", or "medprop"}
}
\value{
plot log risk ratios
}
\description{
Plotting function for survival mediation modeling
}
