% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_seq_pp_eos}
\alias{plot_seq_pp_eos}
\title{Plot end of study posterior predictive trial summaries for sequence of analysis times}
\usage{
plot_seq_pp_eos(dat, df, alternative = "two.sided")
}
\arguments{
\item{dat}{data.frame of class mpti_weib_dat}

\item{df}{object of class pp_eos_df}

\item{alternative}{A data.frame}
}
\value{
ggplot2 object
}
\description{
Plot end of study posterior predictive trial summaries for sequence of analysis times
}
