% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pp_to_df}
\alias{pp_to_df}
\title{Organize posterior predictive distr as data.frame}
\usage{
pp_to_df(x, tp)
}
\arguments{
\item{x}{object of class pp_eos}

\item{tp}{trial time point}
}
\value{
data.frame
}
\description{
Organize posterior predictive distr as data.frame
}
