% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart_tool.R
\name{predict_path}
\alias{predict_path}
\title{predict subtype and leaf assignments}
\usage{
predict_path(tree, newdata, newvarname = "subtype")
}
\arguments{
\item{tree}{A decision tree from rpart object}

\item{newdata}{data.frame}
}
\description{
predict subtype and leaf assignments
}
