% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{riskratio.imstate_data}
\alias{riskratio.imstate_data}
\title{Estimate survival risk ratios and mediation proportion for binary mediator using Cox proportional hazards}
\usage{
\method{riskratio}{imstate_data}(x, tseq = NULL)
}
\arguments{
\item{x}{imstate_data object}

\item{tseq}{times to evaluate risk ratios}
}
\value{
list of risk ratio calculations at observed event times
}
\description{
Estimate survival risk ratios and mediation proportion for binary mediator using Cox proportional hazards
}
