% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{riskratio.mpti_weib_dat}
\alias{riskratio.mpti_weib_dat}
\title{Estimate survival risk ratios and mediation proportion for binary mediator using Cox proportional hazards}
\usage{
\method{riskratio}{mpti_weib_dat}(
  x,
  y = "OS",
  event = "OSevent",
  trt = "trt",
  mediator = "resp",
  covars = NULL,
  tseq = NULL
)
}
\arguments{
\item{x}{data frame}

\item{y}{name of TTF duration}

\item{event}{name of TTF event}

\item{trt}{treatment name}

\item{mediator}{mediator var name (surrogate)}

\item{covars}{names of prognostic covariates}

\item{tseq}{times to evaluate risk ratios}
}
\value{
list of risk ratio calculations at observed event times
}
\description{
Estimate survival risk ratios and mediation proportion for binary mediator using Cox proportional hazards
}
