% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{riskratio.mpti_weib_parm}
\alias{riskratio.mpti_weib_parm}
\title{Calculate true survival risk ratios from MPTI Weibull multi-state mediation model}
\usage{
\method{riskratio}{mpti_weib_parm}(x, maxt = 30, tseq = NULL, npart = 300, nsims = 1000)
}
\arguments{
\item{x}{MPTI Weibull model parameters}

\item{maxt}{time duration upper bound}

\item{tseq}{time grid}

\item{npart}{length out of time partition}

\item{nsims}{number of iterations for simulating response model}
}
\value{
list of risk ratio calculations over time grid
}
\description{
Calculate true survival risk ratios from MPTI Weibull multi-state mediation model
}
