% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio.R
\name{riskratio}
\alias{riskratio}
\title{riskratio}
\usage{
riskratio(x, ...)
}
\arguments{
\item{x}{either data.frame of class mpti_weib_dat or parameter list of class mpti_weib_parm}

\item{...}{additional arguments to class specific function calls}
}
\description{
riskratio
}
