% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_pp_eos}
\alias{summary_pp_eos}
\title{Summarize posterior predictive trial summaries}
\usage{
summary_pp_eos(x, threshold = list(p = 0.05, hr = 0.8))
}
\arguments{
\item{x}{object of class pp_eos}

\item{threshold}{list of thresholds}
}
\value{
list of summaries
}
\description{
Summarize posterior predictive trial summaries
}
