% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{surv_mpti_weib}
\alias{surv_mpti_weib}
\title{Survival function for Weibull MPTI model over interval 0 to t (probability of state 1 or 2)}
\usage{
surv_mpti_weib(parm, t, z, npart = 300)
}
\arguments{
\item{parm}{MPTI Weibull model parameters}

\item{t}{time interval upper bound (0, t)}

\item{z}{vector of model matrix row (treatment, mediatior, prognostic covariates)}

\item{npart}{length out time grid}
}
\value{
vector of survival probabilities over time 0, t
}
\description{
Survival function for Weibull MPTI model over interval 0 to t (probability of state 1 or 2)
}
